+++ src/pages/auth/login/index.ts
@@ -1,0 +1,3 @@
+import LoginPage from "./LoginPage";
+
+export default LoginPage;
